refactor(auth): extract TokenValidator from middleware
