+++ backend-rs/crates/server/src/http/auth/mod.rs
@@ -1,6 +1,7 @@
 pub mod extractor;
 pub mod middleware;
 pub mod roles;
+pub mod validator;
 
 pub use middleware::{AuthLayer, auth_middleware};
 pub use roles::require_any_role;
